Added the URL field and the file name field with the extension
