+++ RunProgram.java
@@ -1,4 +1,4 @@
-// Created oct 01 mon 2018
+// Created oct 03 wed 2018
 // reset && javac -d bin RunProgram.java && java -cp bin RunProgram
 
 public class RunProgram {
